=cdd #17779 Create activator template for Distributed Data
* includes the samples, with tutorial text description
