+++ akka-samples/akka-sample-distributed-data-scala/activator.properties
@@ -1,0 +1,7 @@
+name=akka-sample-distributed-data-scala
+title=Akka Distributed Data Samples with Scala
+description=Akka Distributed Data Samples with Scala
+tags=akka,cluster,scala,sample,distributed-data
+authorName=Akka Team
+authorLink=http://akka.io/
+sourceLink=https://github.com/akka/akka
