+++ akka-samples/akka-sample-distributed-data-scala/project/plugins.sbt
@@ -1,0 +1,4 @@
+
+resolvers += Classpaths.typesafeResolver
+
+addSbtPlugin("com.typesafe.sbt" % "sbt-multi-jvm" % "0.3.8")
