+++ akka-samples/akka-sample-distributed-data-scala/project/sbt-ui.sbt
@@ -1,0 +1,3 @@
+// This plugin represents functionality that is to be added to sbt in the future
+
+addSbtPlugin("org.scala-sbt" % "sbt-core-next" % "0.1.1")
