+++ akka-samples/akka-sample-distributed-data-scala/src/main/resources/application.conf
@@ -1,0 +1,21 @@
+akka {
+  actor {
+    provider = "akka.cluster.ClusterActorRefProvider"
+  }
+  remote {
+    log-remote-lifecycle-events = off
+    netty.tcp {
+      hostname = "127.0.0.1"
+      port = 0
+    }
+  }
+
+  cluster {
+    seed-nodes = [
+      "akka.tcp://ClusterSystem@127.0.0.1:2551",
+      "akka.tcp://ClusterSystem@127.0.0.1:2552"]
+
+    auto-down-unreachable-after = 10s
+  }
+}
+
