+++ akka-samples/akka-sample-distributed-data-scala/src/main/scala/sample/distributeddata/ShoppingCart.scala
@@ -1,7 +1,4 @@
-/**
- * Copyright (C) 2009-2015 Typesafe Inc. <http://www.typesafe.com>
- */
-package sample.datareplication
+package sample.distributeddata
 
 import scala.concurrent.duration._
 import akka.actor.Actor
@@ -10,28 +7,7 @@ import akka.actor.Props
 import akka.cluster.Cluster
 import akka.cluster.ddata.DistributedData
 import akka.cluster.ddata.LWWMap
-import akka.cluster.ddata.Replicator.GetReplicaCount
-import akka.cluster.ddata.Replicator.ReplicaCount
-import akka.cluster.ddata.STMultiNodeSpec
-import akka.remote.testconductor.RoleName
-import akka.remote.testkit.MultiNodeConfig
-import akka.remote.testkit.MultiNodeSpec
-import akka.testkit._
-import com.typesafe.config.ConfigFactory
 import akka.cluster.ddata.LWWMapKey
-
-object ReplicatedShoppingCartSpec extends MultiNodeConfig {
-  val node1 = role("node-1")
-  val node2 = role("node-2")
-  val node3 = role("node-3")
-
-  commonConfig(ConfigFactory.parseString("""
-    akka.loglevel = INFO
-    akka.actor.provider = "akka.cluster.ClusterActorRefProvider"
-    akka.log-dead-letters-during-shutdown = off
-    """))
-
-}
 
 object ShoppingCart {
   import akka.cluster.ddata.Replicator._
@@ -139,76 +115,3 @@ class ShoppingCart(userId: String) extends Actor {
     }
 
 }
-
-class ReplicatedShoppingCartSpecMultiJvmNode1 extends ReplicatedShoppingCartSpec
-class ReplicatedShoppingCartSpecMultiJvmNode2 extends ReplicatedShoppingCartSpec
-class ReplicatedShoppingCartSpecMultiJvmNode3 extends ReplicatedShoppingCartSpec
-
-class ReplicatedShoppingCartSpec extends MultiNodeSpec(ReplicatedShoppingCartSpec) with STMultiNodeSpec with ImplicitSender {
-  import ReplicatedShoppingCartSpec._
-  import ShoppingCart._
-
-  override def initialParticipants = roles.size
-
-  val cluster = Cluster(system)
-  val shoppingCart = system.actorOf(ShoppingCart.props("user-1"))
-
-  def join(from: RoleName, to: RoleName): Unit = {
-    runOn(from) {
-      cluster join node(to).address
-    }
-    enterBarrier(from.name + "-joined")
-  }
-
-  "Demo of a replicated shopping cart" must {
-    "join cluster" in within(10.seconds) {
-      join(node1, node1)
-      join(node2, node1)
-      join(node3, node1)
-
-      awaitAssert {
-        DistributedData(system).replicator ! GetReplicaCount
-        expectMsg(ReplicaCount(roles.size))
-      }
-      enterBarrier("after-1")
-    }
-
-    "handle updates directly after start" in within(15.seconds) {
-      runOn(node2) {
-        shoppingCart ! ShoppingCart.AddItem(LineItem("1", "Apples", quantity = 2))
-        shoppingCart ! ShoppingCart.AddItem(LineItem("2", "Oranges", quantity = 3))
-      }
-      enterBarrier("updates-done")
-
-      awaitAssert {
-        shoppingCart ! ShoppingCart.GetCart
-        val cart = expectMsgType[Cart]
-        cart.items should be(Set(LineItem("1", "Apples", quantity = 2), LineItem("2", "Oranges", quantity = 3)))
-      }
-
-      enterBarrier("after-2")
-    }
-
-    "handle updates from different nodes" in within(5.seconds) {
-      runOn(node2) {
-        shoppingCart ! ShoppingCart.AddItem(LineItem("1", "Apples", quantity = 5))
-        shoppingCart ! ShoppingCart.RemoveItem("2")
-      }
-      runOn(node3) {
-        shoppingCart ! ShoppingCart.AddItem(LineItem("3", "Bananas", quantity = 4))
-      }
-      enterBarrier("updates-done")
-
-      awaitAssert {
-        shoppingCart ! ShoppingCart.GetCart
-        val cart = expectMsgType[Cart]
-        cart.items should be(Set(LineItem("1", "Apples", quantity = 7), LineItem("3", "Bananas", quantity = 4)))
-      }
-
-      enterBarrier("after-3")
-    }
-
-  }
-
-}
-
