+++ project/AkkaBuild.scala
@@ -210,7 +210,8 @@ object AkkaBuild extends Build {
       Seq(sampleCamelJava, sampleCamelScala, sampleClusterJava, sampleClusterScala, sampleFsmScala, sampleFsmJavaLambda,
         sampleMainJava, sampleMainScala, sampleMainJavaLambda, sampleMultiNodeScala,
         samplePersistenceJava, samplePersistenceScala, samplePersistenceJavaLambda,
-        sampleRemoteJava, sampleRemoteScala, sampleSupervisionJavaLambda)
+        sampleRemoteJava, sampleRemoteScala, sampleSupervisionJavaLambda,
+        sampleDistributedDataScala)
   )
 
   lazy val sampleCamelJava = Sample.project("akka-sample-camel-java")
@@ -236,6 +237,8 @@ object AkkaBuild extends Build {
   lazy val sampleRemoteScala = Sample.project("akka-sample-remote-scala")
 
   lazy val sampleSupervisionJavaLambda = Sample.project("akka-sample-supervision-java-lambda")
+  
+  lazy val sampleDistributedDataScala = Sample.project("akka-sample-distributed-data-scala")
 
   lazy val osgiDiningHakkersSampleMavenTest = Project(id = "akka-sample-osgi-dining-hakkers-maven-test",
     base = file("akka-samples/akka-sample-osgi-dining-hakkers-maven-test"),
